+++ code/headers/frame_types.hpp
@@ -1,119 +1,10 @@
-
 #pragma once
 
-#include <cstddef>
-#include <cstdint>
-#include <type_traits>
-
+#include "frames/macros.hpp"
+#include "frames/definitions.hpp"
 #include "frame_enums.hpp"
 
-/*
- * This set of macros is used to quickly add packet helper structs
- * that help the communication system understand how to use
- * and refer to packets. It also checks whether the size of the frame
- * struct is within the hard limits of the protocol.
- * 
- * All frames should be 248 bytes or less, the frame_external_s 
- * is an exception. It is possible there will be another exception for 
- * the robos instructions later.
- */
-#define R2D2_STRINGIFY(x) #x
-#define R2D2_TO_STRING(x) R2D2_STRINGIFY(x)
-
-#define R2D2_OPTIMISE_STRING(Type, MemberName) \
-    template<> \
-    struct supports_string_optimisation<Type> : std::true_type {}; \
-    \
-    template<> \
-    struct string_member_offset<Type> { \
-        constexpr static uint16_t offset = offsetof(Type, MemberName); \
-    };
-
-#define R2D2_OPTIMISE_ARRAY(Type, LengthName, MemberName) \
-    template<> \
-    struct supports_array_optimisation<Type> : std::true_type {}; \
-    \
-    template<> \
-    struct array_member_offset<Type> { \
-        constexpr static uint8_t array_offset = offsetof(Type, MemberName); \
-        constexpr static uint8_t length_offset = offsetof(Type, LengthName); \
-\
-        using array_type = std::remove_pointer_t< \
-            std::decay_t< \
-                decltype(Type::MemberName) \
-            > \
-        >; \
-    }; \
-    static_assert( \
-        std::is_same_v<decltype(Type::LengthName), uint8_t>, \
-        "The length variable of a array optimised frame should be a uint8_t." \
-    );
-
-/**
- * Poisioning is used to prevent people from
- * using structs that are meant purely for the
- * Python bus
- */ 
-#define R2D2_POISON_TYPE(Type) _Pragma(R2D2_TO_STRING(GCC poison Type))
-
-#define R2D2_INTERNAL_FRAME_HELPER(Type, EnumVal, ...) \
-    template<> \
-    struct frame_type_s<Type> { \
-        constexpr static frame_id type = frame_type::EnumVal; \
-    }; \
-    \
-    template<> \
-    struct frame_data_s<frame_type::EnumVal> { \
-        using type = Type; \
-    }; \
-    \
-    static_assert( \
-        std::is_same_v<Type, frame_external_s> \
-        || sizeof(Type) <= 248, "The size of a frame type should not exceed 248 bytes!" \
-    ); \
-    \
-    __VA_ARGS__
-
-/**
- * Travis doesn't like #pragma pack(1), this define makes
- * it so packing is only done on an ARM target.
- */ 
-#if defined(__arm__) || defined(__thumb__)
-#define R2D2_PACK_STRUCT _Pragma("pack(1)")
-#else
-// Empty define
-#define R2D2_PACK_STRUCT
-#endif
-
-// Tag that indicates that the frame is meant for Python only
-#define R2D2_PYTHON_FRAME
-
 namespace r2d2 {
-    /**
-     * The underlying type used for packet
-     * identifiers.
-     */
-    using frame_id = uint8_t;
-
-    /**
-     * Central definition that is used to
-     * determine whether the type T is suitable as
-     * a packet type.
-     *
-     * @tparam T
-     */
-    template<typename T>
-    constexpr bool is_suitable_frame_v = std::is_pod_v<T>;
-
-    /**
-     * Is the given packet type
-     * extended (spans multiple network layer packets)?
-     *
-     * @tparam T
-     */
-    template<typename T>
-    constexpr bool is_extended_frame_v = sizeof(T) > 8;
-
     /**
      * This enum will contain all packet types
      * in the system.
@@ -188,74 +79,6 @@ namespace r2d2 {
      */
     template<typename T>
     constexpr frame_id frame_type_v = frame_type_s<T>::type;
-
-    /**
-     * This struct is specialized to indicate that the
-     * type it is specialized for support the string optimzation.
-     * 
-     * @tparam T
-     */ 
-    template<typename T>
-    struct supports_string_optimisation : std::false_type {};
-
-    /**
-     * Struct that stores the offset of
-     * the string member that can be optimised against.
-     * 
-     * @tparam T
-     */ 
-    template<typename T>
-    struct string_member_offset {
-        constexpr static uint16_t offset = 0;
-    };
-
-    /**
-     * Helper accessor to check for string
-     * optimisation support on the given type.
-     *
-     * @tparam T
-     */  
-    template<typename T>
-    constexpr bool supports_string_optimisation_v = supports_string_optimisation<T>::value;
-
-    /**
-     * Helper accessor to get the string member
-     * offset for the given type.
-     * 
-     * @tparam T
-     */ 
-    template<typename T>
-    constexpr bool string_member_offset_v = string_member_offset<T>::offset;
-
-    /**
-     * This struct is specialized to indicate that the
-     * type it is specialized for support the array optimzation.
-     * 
-     * @tparam T
-     */ 
-    template<typename T>
-    struct supports_array_optimisation : std::false_type {};
-
-    /**
-     * Struct that stores the offset of
-     * the array members that can be optimised against.
-     * 
-     * @tparam T
-     */ 
-    template<typename T>
-    struct array_member_offset {
-        constexpr static uint8_t array_offset = 0;
-        constexpr static uint8_t length = 0;
-    };
-
-    /**
-     * Helper accessor to check for array
-     * optimisation support on the given type.
-     *
-     * @tparam T
-     */  
-    template<typename T>
-    constexpr bool supports_array_optimisation_v = supports_array_optimisation<T>::value;
 
     /**
     * A struct that helps to describe
